+++ src/components/@shared/index.ts
@@ -1,0 +1,7 @@
+export { default as Button } from './Button/styles';
+export { default as CheckBox } from './CheckBox/styles';
+export { default as Empty } from './Empty';
+export { default as Form } from './Form';
+export { default as Input } from './Input';
+export { default as Loader } from './Loader';
+export { default as SnackBar } from './SnackBar';
