+++ src/pages/Login/index.tsx
@@ -1,8 +1,7 @@
 import { useNavigate } from 'react-router-dom';
 import routes from '../../routes';
 
-import Form from '../../components/@shared/Form';
-import Input from '../../components/@shared/Input';
+import { Button, Form, Input } from '../../components/@shared';
 import PageLayout from '../../components/PageLayout';
 
 function Login() {
