+++ src/pages/ProductDetail/index.tsx
@@ -1,4 +1,4 @@
-import Empty from '../../components/@shared/Empty';
+import { Empty } from '../../components/@shared';
 
 function ProductDetail() {
   return <Empty />;
