+++ src/pages/ProductList/index.tsx
@@ -2,8 +2,8 @@ import { useEffect } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { loadProductsAPI, ProductState, selectProductState } from '../../redux/modules/products';
 
+import { Loader } from '../../components/@shared';
 import Product from '../../components/Product';
-import Loader from '../../components/@shared/Loader';
 import { ProductListContainer } from './styles';
 
 function ProductList() {
